Bug #1448699 power: supply: Fix two alerts for code scans
[root cause  ] Fix two alerts for code scans
[changes     ]  charger-manager.c
[side effects] no
[self test   ]: pass
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]: sp7731e_1h10_go2g
[test case]:poweron test
[reviewers   ] jinfeng/yongzhi
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

Code scanning for unsigned and signed types does not perform operations,
this change converts the corresponding int to an unsigned int.

+++ drivers/power/supply/charger-manager.c
@@ -284,8 +284,7 @@ static int cm_init_cap_remap_table(struct charger_desc *desc,
 		dev_err(dev, "%s  get cm-cap-remap-table fail\n", __func__);
 		return 0;
 	}
-
-	desc->cap_remap_table_len = size / (3 * sizeof(__be32));
+	desc->cap_remap_table_len = (u32)size / (3 * sizeof(__be32));
 	desc->cap_remap_table = devm_kzalloc(dev, sizeof(struct cap_remap_table) *
 				(desc->cap_remap_table_len + 1), GFP_KERNEL);
 	if (!desc->cap_remap_table) {
@@ -3972,7 +3971,7 @@ static void cm_track_capacity_monitor(struct charger_manager *cm)
 			 */
 			capacity = (clbcnt - cm->track.start_clbcnt) / 1000;
 			capacity =
-				(total_cap * cm->track.start_cap) / 1000 + capacity;
+				(total_cap * cm->track.start_cap) / 1000 + (u32)capacity;
 
 			if (abs(capacity - total_cap) < total_cap / 2) {
 				set_batt_total_cap(cm, capacity);
